Crate pallet_messenger

Pallet messenger used to communicate between domains and other blockchains.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_messenger

Structs

InitiateChannelParams

Parameters for a new channel between two chains.

Pallet

Pallet messenger used to communicate between chains and other blockchains.

ValidatedRelayMessage

A validated relay message.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

ChainAllowlistUpdate

Parameter to update chain allow list.

Error

pallet-messenger errors

Event

pallet-messenger events

OutboxMessageResult

Traits

Config

Configuration trait of this pallet.

HoldIdentifier

Hold identifier trait for messenger specific balance holds

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.