Module pallet_messenger::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

close_channel

An open channel is closed with a foreign chain. Channel is set to Closed and do not accept or receive any messages.

initiate_channel

A new Channel is initiated with a foreign chain. Next Channel ID is used to assign the new channel. Channel is set to initiated and do not accept or receive any messages.

initiate_domain_update_chain_allowlist

A call to initiate chain allowlist update on domains

relay_message

Receives an Inbox message that needs to be validated and processed.

relay_message_response

Receives a response from the dst_chain for a message in Outbox.

update_consensus_chain_allowlist

A call to update consensus chain allow list.

update_domain_allowlist

An inherent call to update allowlist for domain.