Function pallet_messenger::dispatchables::initiate_channel

pub fn initiate_channel<T: Config>(
    dst_chain_id: ChainId,
    params: InitiateChannelParams,
)

A new Channel is initiated with a foreign chain. Next Channel ID is used to assign the new channel. Channel is set to initiated and do not accept or receive any messages.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::initiate_channel.