Trait pallet_messenger::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type DomainId: Parameter + Member + Default + Copy + MaxEncodedLen;
    type SelfDomainId: Get<Self::DomainId>;
    type DomainTracker: DomainTrackerT<Self::DomainId, <<Self as Config>::Hashing as Hash>::Output>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type MaximumRelayers: Get<u32>;
    type RelayerDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;

    fn get_endpoint_response_handler(
        endpoint: &Endpoint
    ) -> Option<Box<dyn EndpointHandler<Self::DomainId, MessageId>>>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type DomainId: Parameter + Member + Default + Copy + MaxEncodedLen

Domain ID uniquely identifies a Domain.

type SelfDomainId: Get<Self::DomainId>

Gets the domain_id that is treated as src_domain for outgoing messages.

type DomainTracker: DomainTrackerT<Self::DomainId, <<Self as Config>::Hashing as Hash>::Output>

System domain tracker.

type Currency: ReservableCurrency<Self::AccountId>

Currency type pallet uses for fees and deposits.

type MaximumRelayers: Get<u32>

Maximum number of relayers that can join this domain.

type RelayerDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Relayer deposit to become a relayer for this Domain.

Required Methods

fn get_endpoint_response_handler(
endpoint: &Endpoint
) -> Option<Box<dyn EndpointHandler<Self::DomainId, MessageId>>>

function to fetch endpoint response handler by Endpoint.