Crate pallet_messenger

Pallet messenger used to communicate between domains and other blockchains.

Re-exports

pub use weights::WeightInfo;

pub use crate::extensions::weights::WeightInfo as ExtensionWeightInfo;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

extensions

Extensions for unsigned general extrinsics

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_messenger

Structs

EnsureMessengerOrigin

Ensure the messenger origin.

Pallet

Pallet messenger used to communicate between chains and other blockchains.

ValidatedRelayMessage

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

ChainAllowlistUpdate

Parameter to update chain allow list.

Error

pallet-messenger errors

Event

pallet-messenger events

OutboxMessageResult

RawOrigin

Custom origin for validated unsigned extrinsics.

Traits

Config

Configuration trait of this pallet.

HoldIdentifier

Hold identifier trait for messenger specific balance holds

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Origin