Module pallet_messenger::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

ChainAllowlist

An allowlist of chains that can open channel with this chain.

Channels

Stores channel config between two chains. Key points to the foreign chain wrt own chain’s storage name space

DomainChainAllowlistUpdate

A storage to store any allowlist updates to domain. The updates will be cleared in the next block once the previous block has a domain bundle, but a empty value should be left because in the invalid extrinsic root fraud proof the prover need to generate a proof-of-empty-value for the domain.

Inbox

A temporary storage for storing decoded inbox message between pre_dispatch_relay_message and relay_message.

InboxFee

A temporary storage of fees for executing an inbox message. The storage is cleared when the acknowledgement of inbox response is received from the src_chain.

InboxResponses

Stores the message responses of the incoming processed responses. Used by the dst_chains to verify the message response.

MessageWeightTags

Storage to store the weight tags for all the outbox and inbox response messages.

NextChannelId

Stores the next channel id for a foreign chain.

Outbox

Stores the outgoing messages that are awaiting message responses from the dst_chain. Messages are processed in the outbox nonce order of chain’s channel.

OutboxFee

A temporary storage of fees for executing an outbox message and its response from dst_chain. The storage is cleared when src_chain receives the response from dst_chain.

OutboxResponses

A temporary storage for storing decoded outbox response message between pre_dispatch_relay_message_response and relay_message_response.

UpdatedChannels

Temporary storage to store the updated channels between this chain and other chain. Storage is cleared on block initialization.