pallet_messenger

Trait Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type SelfChainId: Get<ChainId>;
    type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId> + Balanced<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type WeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type AdjustedWeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type FeeMultiplier: Get<u32>;
    type OnXDMRewards: OnXDMRewards<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type MmrHash: Parameter + Member + Default + Clone;
    type MmrProofVerifier: MmrProofVerifier<Self::MmrHash, BlockNumberFor<Self>, <<Self as Config>::Hashing as Hash>::Output>;
    type StorageKeys: StorageKeys;
    type DomainOwner: DomainOwner<Self::AccountId>;
    type HoldIdentifier: HoldIdentifier<Self>;
    type ChannelReserveFee: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type ChannelInitReservePortion: Get<Perbill>;
    type DomainRegistration: DomainRegistration;
    type MaxOutgoingMessages: Get<u32>;
    type MessengerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>;
    type NoteChainTransfer: NoteChainTransfer<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type ExtensionWeightInfo: ExtensionWeightInfo;

    // Required method
    fn get_endpoint_handler(
        endpoint: &Endpoint,
    ) -> Option<Box<dyn EndpointHandler<MessageId>>>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type SelfChainId: Get<ChainId>

Gets the chain_id that is treated as src_chain_id for outgoing messages.

type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId> + Balanced<Self::AccountId>

Currency type pallet uses for fees and deposits.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type WeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Weight to fee conversion.

type AdjustedWeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Adjusted Weight to fee conversion. This includes the TransactionPayment Multiper at the time of fee deduction.

type FeeMultiplier: Get<u32>

Fee Multiper for XDM Final fee calculated will fee_multiplier * adjusted_weight_to_fee.

type OnXDMRewards: OnXDMRewards<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Handle XDM rewards.

type MmrHash: Parameter + Member + Default + Clone

Hash type of MMR

type MmrProofVerifier: MmrProofVerifier<Self::MmrHash, BlockNumberFor<Self>, <<Self as Config>::Hashing as Hash>::Output>

MMR proof verifier

type StorageKeys: StorageKeys

Storage key provider.

type DomainOwner: DomainOwner<Self::AccountId>

Domain owner provider.

type HoldIdentifier: HoldIdentifier<Self>

A variation of the Identifier used for holding the funds used for Messenger

type ChannelReserveFee: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Channel reserve fee to open a channel.

type ChannelInitReservePortion: Get<Perbill>

Portion of Channel reserve taken by the protocol if the channel is in init state and is requested to be closed.

type DomainRegistration: DomainRegistration

Type to check if a given domain is registered on Consensus chain.

type MaxOutgoingMessages: Get<u32>

Maximum outgoing messages from a given channel

type MessengerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>

Origin for messenger call.

type NoteChainTransfer: NoteChainTransfer<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Helper to note cross chain XDM fee transfer

type ExtensionWeightInfo: ExtensionWeightInfo

Weight info for extensions

Required Methods

fn get_endpoint_handler( endpoint: &Endpoint, ) -> Option<Box<dyn EndpointHandler<MessageId>>>

function to fetch endpoint response handler by Endpoint.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.