Trait pallet_messenger::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type SelfChainId: Get<ChainId>;
    type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type WeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type OnXDMRewards: OnXDMRewards<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type MmrHash: Parameter + Member + Default + Clone;
    type MmrProofVerifier: MmrProofVerifier<Self::MmrHash, BlockNumberFor<Self>, <<Self as Config>::Hashing as Hash>::Output>;
    type StorageKeys: StorageKeys;
    type DomainOwner: DomainOwner<Self::AccountId>;
    type HoldIdentifier: HoldIdentifier<Self>;
    type ChannelReserveFee: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;

    // Required method
    fn get_endpoint_handler(
        endpoint: &Endpoint
    ) -> Option<Box<dyn EndpointHandler<MessageId>>>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type SelfChainId: Get<ChainId>

Gets the chain_id that is treated as src_chain_id for outgoing messages.

type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId>

Currency type pallet uses for fees and deposits.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type WeightToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Weight to fee conversion.

type OnXDMRewards: OnXDMRewards<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Handle XDM rewards.

type MmrHash: Parameter + Member + Default + Clone

Hash type of MMR

type MmrProofVerifier: MmrProofVerifier<Self::MmrHash, BlockNumberFor<Self>, <<Self as Config>::Hashing as Hash>::Output>

MMR proof verifier

type StorageKeys: StorageKeys

Storage key provider.

type DomainOwner: DomainOwner<Self::AccountId>

Domain owner provider.

type HoldIdentifier: HoldIdentifier<Self>

A variation of the Identifier used for holding the funds used for Messenger

type ChannelReserveFee: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Channel reserve fee to open a channel.

Required Methods

fn get_endpoint_handler( endpoint: &Endpoint ) -> Option<Box<dyn EndpointHandler<MessageId>>>

function to fetch endpoint response handler by Endpoint.

Object Safety

This trait is not object safe.